Initial scanner implementation
The scanner's job is to split the input data into lines and handle the
indents and dedents, producing synthetic tokens so that the parser can
treat indentation changes as a kind of "punctuation".

Due to their unusual interaction with the indentation logic the scanner
must also distinguish lines that are part of "literal" blocks, and return
them as a separate token type so that the parser can handle them in a
special way.