Allow the parser to "put back" a token with a prefix removed
When we parse structures that are introduced by some sort of marker,
like a bullet list, the first token of that structure includes both the
marker and the first line of the contents.

To allow the two to be parsed by separate parser states, the scanner now
allows the parser to "push back" a token with the prefix removed, which
means the next read will obtain that token and be able to process it in
whatever way makes sense for the new parser state.