Structure elements have positions
While it feels not worth the trouble to attach positions to *all* of the
DOM nodes, having positions on the sections is likely to prove useful
since the sections provide the top-level structure of the document, and
in a multi-file composition (e.g. a book) it could be useful to get
a summary of how the sections are distributed across all the files.