When emitting LATE_INDENT, don't also emit DEDENT
A LATE_INDENT is a signal to the *current* structure that it needs to
turn everything it's seen so far into a nested indent level, so we don't
want to DEDENT first or else we'll exit from the current structure early.

Instead, we emit LATE_INDENT *instead of* DEDENT, allowing the current
structure to deal with it and continue, and then DEDENT is eventually
emitted once we return to the previous indent level, thus finally exiting
the current structure.