Rearrange how we deal with structural model parsing
We want to share as much code as possible between body-only parsing and
structure model parsing, so we create here a new structureModelParser
struct that acts as an event-driven parser for block and structural
elements.

The different parsing modes can then respond differently to these events
in order to deal with the minor differences in processing between these
cases.