Blockquote attributions
The attribution syntax requires some trickiness in the parsing since RST
permits a sequence of different blockquotes with the attribution markers
as delimiters, without any intervening dedent/indent sequence.

We introduce here a divergence from the RST spec and the docutils
implementation, where we can't parse a blockquote with attribution
followed by another blockquote that is indented to the same or greater
level as the attribution text.